BUGFIX: IBus.Subscribe[Async] method throws an error when the queue is previously created on RabbitMQ with different args than the default
When subscribing to an exchange using the IBus.Subscribe[Async] method
both queue and exchange are declared. If the queue was previously
created on RabbitMQ using arguments different than the default values on
IAdvancedBus.QueueDeclare[Async] method, the RabbitMQ will thrown the
following error:
The AMQP operation was interrupted: AMQP close-reason, initiated by
Peer, code=406, text="PRECONDITION_FAILED - inequivalent arg
'x-max-priority' for queue 'my_queue' in vhost '/': received none but
current is the value '9' of type 'long' [...]

To solve it, the user should be able to inject the arguments of the
queue declaration in the IBus.Subscribe[Async] method call, so the
IAdvancedBus.QueueDeclare[Async] declaration can match the previously
created queue. The existing argument Action<ISubscriptionConfiguration>
was used as inspiration to do the same changing the exchange and queue
args declarations.

+++ Source/EasyNetQ/RabbitAdvancedBus.cs
@@ -98,7 +98,7 @@ namespace EasyNetQ
             connection.Initialize();
         }
 
-        
+
         // ---------------------------------- consume --------------------------------------
 
         public IDisposable Consume<T>(IQueue queue, Action<IMessage<T>, MessageReceivedInfo> onMessage) where T : class
@@ -180,11 +180,11 @@ namespace EasyNetQ
         // -------------------------------- publish ---------------------------------------------
 
         public void Publish(
-            IExchange exchange, 
-            string routingKey, 
-            bool mandatory, 
+            IExchange exchange,
+            string routingKey,
+            bool mandatory,
             bool immediate,
-            MessageProperties messageProperties, 
+            MessageProperties messageProperties,
             byte[] body)
         {
             try
@@ -198,10 +198,10 @@ namespace EasyNetQ
         }
 
         public void Publish<T>(
-            IExchange exchange, 
-            string routingKey, 
-            bool mandatory, 
-            bool immediate, 
+            IExchange exchange,
+            string routingKey,
+            bool mandatory,
+            bool immediate,
             IMessage<T> message) where T : class
         {
             try
@@ -276,31 +276,31 @@ namespace EasyNetQ
         // ---------------------------------- Exchange / Queue / Binding -----------------------------------
 
         public virtual IQueue QueueDeclare(
-            string name, 
-            bool passive = false, 
-            bool durable = true, 
-            bool exclusive = false, 
+            string name,
+            bool passive = false,
+            bool durable = true,
+            bool exclusive = false,
             bool autoDelete = false,
-            int? perQueueMessageTtl  = null,
+            int? perQueueMessageTtl = null,
             int? expires = null,
             byte? maxPriority = null,
-            string deadLetterExchange = null, 
+            string deadLetterExchange = null,
             string deadLetterRoutingKey = null)
         {
-            return QueueDeclareAsync(name, passive, durable, exclusive, autoDelete, perQueueMessageTtl, 
+            return QueueDeclareAsync(name, passive, durable, exclusive, autoDelete, perQueueMessageTtl,
                 expires, maxPriority, deadLetterExchange, deadLetterRoutingKey).Result;
         }
 
         public Task<IQueue> QueueDeclareAsync(
-            string name, 
-            bool passive = false, 
-            bool durable = true, 
-            bool exclusive = false, 
+            string name,
+            bool passive = false,
+            bool durable = true,
+            bool exclusive = false,
             bool autoDelete = false,
-            int? perQueueMessageTtl  = null,
+            int? perQueueMessageTtl = null,
             int? expires = null,
             byte? maxPriority = null,
-            string deadLetterExchange = null, 
+            string deadLetterExchange = null,
             string deadLetterRoutingKey = null)
         {
             Preconditions.CheckNotNull(name, "name");
@@ -322,7 +322,7 @@ namespace EasyNetQ
             }
             if (maxPriority.HasValue)
             {
-                arguments.Add("x-max-priority", maxPriority.Value);
+                arguments.Add("x-max-priority", (int)maxPriority.Value); // cast to Int32 because currently there is no Byte implementation (2015-04-19) -> https://github.com/rabbitmq/rabbitmq-dotnet-client/blob/7c1a7d757dd62c184e8bbfd73288eccf465b4d05/projects/client/RabbitMQ.Client/src/client/impl/WireFormatting.cs
             }
             if (!string.IsNullOrEmpty(deadLetterExchange))
             {
@@ -384,12 +384,12 @@ namespace EasyNetQ
         }
 
         public Task<IExchange> ExchangeDeclareAsync(
-            string name, 
-            string type, 
-            bool passive = false, 
-            bool durable = true, 
-            bool autoDelete = false, 
-            bool @internal = false, 
+            string name,
+            string type,
+            bool passive = false,
+            bool durable = true,
+            bool autoDelete = false,
+            bool @internal = false,
             string alternateExchange = null,
             bool delayed = false)
         {
@@ -424,7 +424,7 @@ namespace EasyNetQ
 
                         return (IExchange)new Exchange(name);
                     });
-       }
+        }
 
         public virtual void ExchangeDelete(IExchange exchange, bool ifUnused = false)
         {
@@ -531,15 +531,15 @@ namespace EasyNetQ
             else
             {
                 var message = messageSerializationStrategy.DeserializeMessage<T>(result.Properties, result.Body);
-                if (message.MessageType == typeof (T))
+                if (message.MessageType == typeof(T))
                 {
                     return new BasicGetResult<T>(message);
                 }
                 else
                 {
-                    logger.ErrorWrite("Incorrect message type returned from Get." + 
+                    logger.ErrorWrite("Incorrect message type returned from Get." +
                         "Expected {0}, but was {1}", typeof(T).Name, message.MessageType.Name);
-                    throw new EasyNetQException("Incorrect message type returned from Get." + 
+                    throw new EasyNetQException("Incorrect message type returned from Get." +
                         "Expected {0}, but was {1}", typeof(T).Name, message.MessageType.Name);
                 }
             }
